fix path issue

+++ install/config/config.sh
@@ -2,5 +2,7 @@
 mkdir -p ~/.config
 cp -R ~/.local/share/macstrap/config/* ~/.config/
 
-# Use default zshrc from macstrap
-cp ~/.local/share/macstrap/default/zshrc ~/.zshrc
+# Use default zshrc from macstrap if it exists
+if [[ -f ~/.local/share/macstrap/default/zshrc ]]; then
+  cp ~/.local/share/macstrap/default/zshrc ~/.zshrc
+fi
